+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 ### Added
 - View in Jamf button to mobile device info box in Computers and Phones
 - When merging items, a GLPI item will be preselected if the UUID/UDID or name matches.
+ - View in Jamf button to mobile device info box in Computers and Phones
+ - When merging items, a GLPI item will be preselected if the UDID/UUID matches (or if the name matches as a backup check)
+ - View and sync Extension Attributes
+
+
+## [1.0.1]
+
+### Fixed
+- Rights on item forms
+- JSS item links
 
 ## [1.0.0]